+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### BUG FIXES
 
+- `[libs/autofile]` unregister the SIGHUP handler in `AutoFile.Close` so a later SIGHUP no longer panics the process with "send on closed channel"
+  ([\#6067](https://github.com/cometbft/cometbft/pull/6067))
 - `[lp2p]` fix flaky MsgBytesFilter test by avoiding broadcast race
   ([\#6053](https://github.com/cometbft/cometbft/pull/6053))
 - `[spec]` fix the inductive invariant `spec/light-client/accountability`